+++ phpunit.xml
@@ -11,7 +11,7 @@
 >
     <testsuites>
         <testsuite name="tests">
-            <directory suffix=".php">tests/</directory>
+            <directory suffix=".php">Tests/</directory>
         </testsuite>
     </testsuites>
     <filter>
